Moving remaining token functions to helpers
