+++ user/helpers_test.go
@@ -1,9 +1,12 @@
 package user
 
 import (
-	"reflect"
-	"testing"
 	"net/http"
+	"reflect"
+	"strconv"
+	"testing"
+
+	"github.com/tidepool-org/shoreline/token"
 )
 
 func Test_FirstStringNotEmpty_None(t *testing.T) {
@@ -144,7 +147,7 @@ func Test_appendUserLoginInProgress_RateLimitExceeded(t *testing.T) {
 	shoreline.loginLimiter.totalInProgress = shoreline.ApiConfig.MaxConcurrentLogin
 	shoreline.loginLimiter.usersInProgress.Init()
 	code, _ := shoreline.appendUserLoginInProgress(user)
-	if(code != http.StatusTooManyRequests) {
+	if code != http.StatusTooManyRequests {
 		t.Fatalf("appendUserLoginInProgress should return an http too many requests error when the limit is execedded")
 	}
 }
@@ -155,7 +158,7 @@ func Test_appendUserLoginInProgress_UserAlreadyLoginIn(t *testing.T) {
 	shoreline.loginLimiter.usersInProgress.Init()
 	code, _ := shoreline.appendUserLoginInProgress(user)
 	code, _ = shoreline.appendUserLoginInProgress(user)
-	if(code != http.StatusTooManyRequests) {
+	if code != http.StatusTooManyRequests {
 		t.Fatalf("appendUserLoginInProgress should return an http too many requests error when the user is in the in progress list")
 	}
 }
@@ -164,13 +167,13 @@ func Test_appendUserLoginInProgress_NoProblem(t *testing.T) {
 	resetLoginLimitConfig()
 	shoreline.loginLimiter.usersInProgress.Init()
 	code, _ := shoreline.appendUserLoginInProgress(user)
-	if(code != http.StatusOK) {
+	if code != http.StatusOK {
 		t.Fatalf("appendUserLoginInProgress should return an http status ok when the user is not in the in progress list")
 	}
-	if(shoreline.loginLimiter.totalInProgress != 1) {
+	if shoreline.loginLimiter.totalInProgress != 1 {
 		t.Fatalf("appendUserLoginInProgress should increment the total login counter")
 	}
-	if(shoreline.loginLimiter.usersInProgress.Len() !=1){
+	if shoreline.loginLimiter.usersInProgress.Len() != 1 {
 		t.Fatalf("appendUserLoginInProgress should add users into userInProgress list")
 	}
 }
@@ -182,28 +185,28 @@ func Test_appendUserLoginInProgress_BlockParallelDisabled(t *testing.T) {
 	shoreline.loginLimiter.totalInProgress = shoreline.ApiConfig.MaxConcurrentLogin
 	shoreline.loginLimiter.usersInProgress.Init()
 	code, _ := shoreline.appendUserLoginInProgress(user)
-	if(code != http.StatusTooManyRequests) {
+	if code != http.StatusTooManyRequests {
 		t.Fatalf("appendUserLoginInProgress should return an http too many requests error when the limit is execedded")
 	}
 	shoreline.loginLimiter.totalInProgress = 0
 	code, _ = shoreline.appendUserLoginInProgress(user)
-	if(code != http.StatusOK) {
+	if code != http.StatusOK {
 		t.Fatalf("appendUserLoginInProgress should return an http status ok when BlockParallelLogin config is set to false")
 	}
-	if(shoreline.loginLimiter.totalInProgress != 1) {
+	if shoreline.loginLimiter.totalInProgress != 1 {
 		t.Fatalf("appendUserLoginInProgress should increment the total login counter")
 	}
-	if(shoreline.loginLimiter.usersInProgress.Len() !=0){
+	if shoreline.loginLimiter.usersInProgress.Len() != 0 {
 		t.Fatalf("appendUserLoginInProgress should leave the userInProgress list empty when BlockParallelLogin config is set to false")
 	}
 	code, _ = shoreline.appendUserLoginInProgress(user)
-	if(code != http.StatusOK) {
+	if code != http.StatusOK {
 		t.Fatalf("appendUserLoginInProgress should return an http status ok when BlockParallelLogin config is set to false")
 	}
-	if(shoreline.loginLimiter.totalInProgress != 2) {
+	if shoreline.loginLimiter.totalInProgress != 2 {
 		t.Fatalf("appendUserLoginInProgress should increment the total login counter")
 	}
-	if(shoreline.loginLimiter.usersInProgress.Len() !=0){
+	if shoreline.loginLimiter.usersInProgress.Len() != 0 {
 		t.Fatalf("appendUserLoginInProgress should leave the userInProgress list empty when BlockParallelLogin config is set to false")
 	}
 }
@@ -214,10 +217,10 @@ func Test_removeUserLoginInProgress(t *testing.T) {
 	shoreline.loginLimiter.usersInProgress.Init()
 	_, elem := shoreline.appendUserLoginInProgress(user)
 	shoreline.removeUserLoginInProgress(elem)
-	if(shoreline.loginLimiter.totalInProgress != 0) {
+	if shoreline.loginLimiter.totalInProgress != 0 {
 		t.Fatalf("removeUserLoginInProgress should decrement the total login counter")
 	}
-	if(shoreline.loginLimiter.usersInProgress.Len() !=0){
+	if shoreline.loginLimiter.usersInProgress.Len() != 0 {
 		t.Fatalf("removeUserLoginInProgress should remove users from userInProgress list")
 	}
 }
@@ -229,20 +232,57 @@ func Test_removeUserLoginInProgress_BlockParallelDisabled(t *testing.T) {
 	shoreline.loginLimiter.usersInProgress.Init()
 	_, elem := shoreline.appendUserLoginInProgress(user)
 	shoreline.removeUserLoginInProgress(elem)
-	if(shoreline.loginLimiter.totalInProgress != 0) {
+	if shoreline.loginLimiter.totalInProgress != 0 {
 		t.Fatalf("removeUserLoginInProgress should decrement the total login counter")
 	}
-	if(shoreline.loginLimiter.usersInProgress.Len() !=0){
+	if shoreline.loginLimiter.usersInProgress.Len() != 0 {
 		t.Fatalf("removeUserLoginInProgress should leave the userInProgress list empty when BlockParallelLogin config is set to false")
 	}
 
 	_, elem = shoreline.appendUserLoginInProgress(user)
-	shoreline.appendUserLoginInProgress( &User{Username: "test4@test.com"})
+	shoreline.appendUserLoginInProgress(&User{Username: "test4@test.com"})
 	shoreline.removeUserLoginInProgress(elem)
-	if(shoreline.loginLimiter.totalInProgress != 1) {
+	if shoreline.loginLimiter.totalInProgress != 1 {
 		t.Fatalf("removeUserLoginInProgress should decrement the total login counter")
 	}
-	if(shoreline.loginLimiter.usersInProgress.Len() !=0){
+	if shoreline.loginLimiter.usersInProgress.Len() != 0 {
 		t.Fatalf("removeUserLoginInProgress should leave the userInProgress list empty when BlockParallelLogin config is set to false")
+	}
+}
+
+func Test_extractTokenDuration(t *testing.T) {
+
+	request, _ := http.NewRequest("GET", "", nil)
+	givenDuration := strconv.FormatFloat(float64(10), 'f', -1, 64)
+
+	request.Header.Add(token.TOKEN_DURATION_KEY, givenDuration)
+
+	duration := extractTokenDuration(request)
+
+	if strconv.FormatInt(duration, 10) != givenDuration {
+		t.Fatalf("Duration should have been set [%s] but was [%s] ", givenDuration, strconv.FormatInt(duration, 10))
+	}
+
+}
+
+func Test_hasServerToken(t *testing.T) {
+	tokenTestData := &token.TokenData{UserId: "2341", IsServer: true, DurationSecs: 1}
+	tokenTestConfig := token.TokenConfig{DurationSecs: 3600, Secret: "my secret"}
+
+	token, _ := token.CreateSessionToken(tokenTestData, tokenTestConfig)
+
+	if hasServerToken(token.ID, tokenTestConfig.Secret) == false {
+		t.Fatal("We should have got a server Token")
+	}
+}
+
+func Test_hasServerToken_false(t *testing.T) {
+	tokenTestData := &token.TokenData{UserId: "2341", IsServer: false, DurationSecs: 1}
+	tokenTestConfig := token.TokenConfig{DurationSecs: 3600, Secret: "my secret"}
+
+	token, _ := token.CreateSessionToken(tokenTestData, tokenTestConfig)
+
+	if hasServerToken(token.ID, tokenTestConfig.Secret) != false {
+		t.Fatal("We should have not got a server Token")
 	}
 }
